+++ Pandora/Paradigm/Basis/Twister.hs
@@ -3,7 +3,7 @@ module Pandora.Paradigm.Basis.Twister (Twister (..), unwrap, coiterate, section)
 import Pandora.Core.Functor (type (:.:), type (~>))
 import Pandora.Core.Morphism ((.))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>), comap))
-import Pandora.Pattern.Functor.Exclusive (Exclusive (exclusive))
+import Pandora.Pattern.Functor.Avoidable (Avoidable (idle))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Extractable (Extractable (extract))
 import Pandora.Pattern.Functor.Alternative (Alternative ((<+>)))
@@ -24,8 +24,8 @@ data Twister t a = a :< (t :.: Twister t) a
 instance Covariant t => Covariant (Twister t) where
 	f <$> (x :< xs) = f x :< ((comap . comap) f xs)
 
-instance Exclusive t => Pointable (Twister t) where
-	point x = x :< exclusive
+instance Avoidable t => Pointable (Twister t) where
+	point x = x :< idle
 
 instance Covariant t => Extractable (Twister t) where
 	extract (x :< _) = x
@@ -43,7 +43,7 @@ instance Alternative t => Bindable (Twister t) where
 instance Covariant t => Extendable (Twister t) where
 	x =>> f = f x :< comap (extend f) (unwrap x)
 
-instance (Exclusive t, Alternative t) => Monad (Twister t) where
+instance (Avoidable t, Alternative t) => Monad (Twister t) where
 
 instance Covariant t => Comonad (Twister t) where