+++ Pandora/Paradigm/Basis/Yoneda.hs
@@ -4,7 +4,7 @@ import Pandora.Core.Morphism ((.), ($), (!), identity)
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>), comap))
 import Pandora.Pattern.Functor.Alternative (Alternative ((<+>)))
 import Pandora.Pattern.Functor.Applicative (Applicative ((<*>)))
-import Pandora.Pattern.Functor.Exclusive (Exclusive (exclusive))
+import Pandora.Pattern.Functor.Avoidable (Avoidable (idle))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Extractable (Extractable (extract))
 import Pandora.Pattern.Functor.Adjoint (Adjoint (phi, psi))
@@ -21,8 +21,8 @@ instance Alternative t => Alternative (Yoneda t) where
 instance Applicative t => Applicative (Yoneda t) where
 	Yoneda f <*> Yoneda x = Yoneda (\g -> f (g .) <*> x identity)
 
-instance Exclusive t => Exclusive (Yoneda t) where
-	exclusive = Yoneda (exclusive !)
+instance Avoidable t => Avoidable (Yoneda t) where
+	idle = Yoneda (idle !)
 
 instance Pointable t => Pointable (Yoneda t) where
 	point x = Yoneda (\f -> point $ f x)